+++ src/main/java/com/osmium/integration/ModMenuApiImpl.java
@@ -1,0 +1,11 @@
+package com.osmium.integration;
+
+import com.terraformersmc.modmenu.api.ConfigScreenFactory;
+import com.terraformersmc.modmenu.api.ModMenuApi;
+
+public class ModMenuApiImpl implements ModMenuApi {
+    @Override
+    public ConfigScreenFactory<?> getModConfigScreenFactory() {
+        return ModMenuIntegration::createConfigScreen;
+    }
+}